+++ lib/card_settings.dart
@@ -20,6 +20,8 @@ export 'widgets/card_settings_time_picker.dart';
 export 'widgets/card_settings_button.dart';
 export 'widgets/card_settings_instructions.dart';
 export 'widgets/card_settings_currency.dart';
+export 'widgets/card_settings_email.dart';
+export 'widgets/card_settings_password.dart';
 
 export 'helpers/picker_dialog.dart';
 export 'helpers/scroll_picker.dart';